src/test/java/cn/cloudx/weichatsell/service/impl/SellerServiceImplTest.java
package cn.cloudx.weichatsell.service.impl;

import cn.cloudx.weichatsell.dataobject.SellerInfo;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import static org.junit.Assert.assertEquals;


@SpringBootTest
@RunWith(SpringRunner.class)
public class SellerServiceImplTest {

    @Autowired
    private SellerServiceImpl sellerService;

    @Test
    public void findSellerInfoByOpenid() {
        SellerInfo abc = sellerService.findSellerInfoByOpenid("abc");
        assertEquals("abc", abc.getOpenid());
    }
}